+++ rayTracer/src/cube.h
@@ -13,7 +13,7 @@ public:
     ~Cube();
 
     virtual void intersectRay(Ray& ray, std::vector<Geometry::Intersection<Shape>>& intersections) override;
-    virtual glm::vec4 normalAt(glm::vec4 point) override;
+    virtual glm::dvec4 normalAt(glm::dvec4 point) override;
     virtual std::string type() override;
 
 private:
