+++ rayTracer/src/cylinder.cpp
@@ -1,12 +1,12 @@
 #include "cylinder.h"
 
 Cylinder::Cylinder() : Shape() {
-  this->minimum = -std::numeric_limits<float>::infinity();
-  this->maximum = std::numeric_limits<float>::infinity();
+  this->minimum = -std::numeric_limits<double>::infinity();
+  this->maximum = std::numeric_limits<double>::infinity();
   this->capped = false;
 }
 
-Cylinder::Cylinder(float minimum, float maximum, bool capped) : Shape() {
+Cylinder::Cylinder(double minimum, double maximum, bool capped) : Shape() {
   this->minimum = minimum;
   this->maximum = maximum;
   this->capped = capped;
@@ -17,31 +17,31 @@ Cylinder::~Cylinder() {}
 void Cylinder::intersectRay(Ray& ray, std::vector<Geometry::Intersection<Shape>>& intersections) { 
     Ray transformedRay = transformRay(ray);
 
-    float a = std::pow(transformedRay.direction.x, 2) +
+    double a = std::pow(transformedRay.direction.x, 2) +
         std::pow(transformedRay.direction.z, 2);
 
     if (std::abs(a) < Geometry::EPSILON)
         return;
 
-    float b = (2 * transformedRay.origin.x * transformedRay.direction.x) +
+    double b = (2 * transformedRay.origin.x * transformedRay.direction.x) +
         (2 * transformedRay.origin.z * transformedRay.direction.z);
 
-    float c = std::pow(transformedRay.origin.x, 2) +
+    double c = std::pow(transformedRay.origin.x, 2) +
         std::pow(transformedRay.origin.z, 2) - 1;
 
-    float discriminant = std::pow(b, 2) - 4 * a * c;
+    double discriminant = std::pow(b, 2) - 4 * a * c;
 
     if (discriminant < 0)
         return;
 
-    float t1 = (-b - std::sqrt(discriminant)) / (2 * a);
-    float t2 = (-b + std::sqrt(discriminant)) / (2 * a);
+    double t1 = (-b - std::sqrt(discriminant)) / (2 * a);
+    double t2 = (-b + std::sqrt(discriminant)) / (2 * a);
 
-    float y1 = transformedRay.origin.y + t1 * transformedRay.direction.y;
+    double y1 = transformedRay.origin.y + t1 * transformedRay.direction.y;
     if (this->minimum < y1 && y1 < this->maximum)
         intersections.push_back(Geometry::Intersection<Shape>{t1, this});
 
-    float y2 = transformedRay.origin.y + t2 * transformedRay.direction.y;
+    double y2 = transformedRay.origin.y + t2 * transformedRay.direction.y;
     if (this->minimum < y2 && y2 < this->maximum)
         intersections.push_back(Geometry::Intersection<Shape>{t2, this});
 
@@ -49,30 +49,30 @@ void Cylinder::intersectRay(Ray& ray, std::vector<Geometry::Intersection<Shape>>
 
 }
 
-glm::vec4 Cylinder::normalAt(glm::vec4 point) {
-  glm::vec4 objectPoint = this->inverseTransform * point;
+glm::dvec4 Cylinder::normalAt(glm::dvec4 point) {
+  glm::dvec4 objectPoint = this->inverseTransform * point;
 
-  float dist = std::pow(objectPoint.x, 2) + std::pow(objectPoint.z, 2);
-  glm::vec4 objectNormal;
+  double dist = std::pow(objectPoint.x, 2) + std::pow(objectPoint.z, 2);
+  glm::dvec4 objectNormal;
 
   if (dist < 1 && objectPoint.y >= (this->maximum - Geometry::EPSILON))
-    objectNormal = glm::vec4(0.f, 1.f, 0.f, 0.f);
+    objectNormal = glm::dvec4(0.0, 1.0, 0.0, 0.0);
   else if (dist < 1 && objectPoint.y <= (this->minimum + Geometry::EPSILON))
-    objectNormal = glm::vec4(0.f, -1.f, 0.f, 0.f);
+    objectNormal = glm::dvec4(0.0, -1.0, 0.0, 0.0);
   else
-    objectNormal = glm::vec4(objectPoint.x, 0.f, objectPoint.z, 0.f);
+    objectNormal = glm::dvec4(objectPoint.x, 0.0, objectPoint.z, 0.0);
 
-  glm::vec4 worldNormal = glm::transpose(this->inverseTransform) * objectNormal;
-  worldNormal.w = 0.f;
+  glm::dvec4 worldNormal = glm::transpose(this->inverseTransform) * objectNormal;
+  worldNormal.w = 0.0;
 
   return glm::normalize(worldNormal);
 }
 
 std::string Cylinder::type() { return "Cylinder"; }
 
-bool Cylinder::checkCap(Ray &ray, float t) {
-  float x = ray.origin.x + t * ray.direction.x;
-  float z = ray.origin.z + t * ray.direction.z;
+bool Cylinder::checkCap(Ray &ray, double t) {
+  double x = ray.origin.x + t * ray.direction.x;
+  double z = ray.origin.z + t * ray.direction.z;
 
   return (std::pow(x, 2) + std::pow(z, 2)) <= 1;
 }
@@ -81,11 +81,11 @@ void Cylinder::intersectCaps(Ray &ray, std::vector<Geometry::Intersection<Shape>
   if (!this->capped || std::abs(ray.direction.y) < Geometry::EPSILON)
     return;
 
-  float t1 = (this->minimum - ray.origin.y) / ray.direction.y;
+  double t1 = (this->minimum - ray.origin.y) / ray.direction.y;
   if (checkCap(ray, t1))
       intersections.push_back(Geometry::Intersection<Shape>{t1, this});
 
-  float t2 = (this->maximum - ray.origin.y) / ray.direction.y;
+  double t2 = (this->maximum - ray.origin.y) / ray.direction.y;
   if (checkCap(ray, t2))
       intersections.push_back(Geometry::Intersection<Shape>{t2, this});
