+++ rayTracer/src/renderer.h
@@ -45,19 +45,19 @@ public:
   static const int RAYS_PER_PIXEL = 1;
   Canvas canvas;
 
-  glm::vec3 colourAt(Ray &ray, World &world, short remaining);
-  glm::vec3 reflectColour(Geometry::Intersection<Shape> *hit, World &world,
+  glm::dvec3 colourAt(Ray &ray, World &world, short remaining);
+  glm::dvec3 reflectColour(Geometry::Intersection<Shape> *hit, World &world,
                           short remaining);
-  glm::vec3 lighting(Shape *shape, std::shared_ptr<PointLight> &light,
-                     glm::vec4 &point, glm::vec4 &eyev, glm::vec4 &normalv,
+  glm::dvec3 lighting(Shape *shape, std::shared_ptr<PointLight> &light,
+                     glm::dvec4 &point, glm::dvec4 &eyev, glm::dvec4 &normalv,
                      bool &inShadow);
-  glm::vec3 shadeHit(Geometry::Intersection<Shape> *hit, World &world,
+  glm::dvec3 shadeHit(Geometry::Intersection<Shape> *hit, World &world,
                      short remaining);
-  glm::vec3 refractedColour(Geometry::Intersection<Shape> *hit, World &world,
+  glm::dvec3 refractedColour(Geometry::Intersection<Shape> *hit, World &world,
                             short remaining);
-  bool isShadowed(glm::vec4 &point, World &world);
+  bool isShadowed(glm::dvec4 &point, World &world);
 
   void render(World &world);
   void renderPixel(World &world, std::pair<int, int> &pixel,
-                   int sqrtRaysPerPixel, float halfSubPixelSize);
+                   int sqrtRaysPerPixel, double halfSubPixelSize);
 };
